requirements.txt
flask
pydantic
pytest
black
pymongo
pyjwt
argon2-cffi
requests
python-dotenv
gunicorn
flask-cors
python-json-logger
